requirements.txt
django-easy-pdf
xhtml12pdf>=0.0.6
reportlab>=2.7,<3